Fix bundler file dependency tracking
